+++ app/globals.css
@@ -19,6 +19,10 @@ a {
   text-decoration: none;
 }
 
+ul {
+  list-style: none;
+}
+
 .mui-o9k5xi-MuiInputBase-root-MuiOutlinedInput-root {
   max-height: 38px;
 }
