--- page/postModify.tsx
@@ -1,5 +1,0 @@
-const PostModifyPage = () => {
-  return <div>PostModifyPage</div>;
-};
-
-export default PostModifyPage;
