+++ public/assets/imgSrc.ts
@@ -1,4 +1,5 @@
 import HamkkeLogo from "@/public/assets/hamkke_logo.png";
 import Error404 from "@/public/assets/errorpage_404.png";
+import ErrorPageImg from "@/public/assets/error.png";
 
-export { HamkkeLogo, Error404 };
+export { HamkkeLogo, Error404, ErrorPageImg };
